fix Cohere models by depending on langchain_cohere
